Fixed encoding of request and decoding of response.

+++ IoTEdgeModbus/Modules/AzureIoTEdgeModbus/Slave/Decoding/IModbusDataDecoder.cs
@@ -14,6 +14,6 @@
         /// </remarks>
         IEnumerable<DecodedValue> GetValues(byte[] bytesToConvert, ReadOperation operation);
 
-        int GetByteCount(int valuesToRead);
+        short GetByteCount(short valuesToRead);
     }
 }
